working training

+++ model/utils.py
@@ -119,7 +119,7 @@ def collate_double(batch:Dict[torch.tensor, str]) -> Tuple[torch.tensor, str]:
     collate function for the ObjectsDataSet.
     Only used by the dataloader.
     """
-    images = [sample['img'].float() for sample in batch]
+    images = [sample['img'].type(torch.double) for sample in batch]
     boxes_labelled = [sample['boxes_labelled'] for sample in batch]
     images_name = [sample['img_file_name'] for sample in batch]
     target_file_name = [sample['target_file_name'] for sample in batch]
